requirements.txt
flask
flask-restx
gunicorn
flask-session
scikit-learn
pandas
spotipy
flask-cors
python-decouple
flake8
nose
coverage